examples/demo_company/requirements.txt
# AI Workforce OS - Demo Company Dependencies

# Core framework
crewai>=0.28.0
crewai-tools>=0.2.0

# LLM providers
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Local models support (optional)
# ollama>=0.1.0

# UI and interaction
gradio>=4.0.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0